api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0

cryptography==41.0.7
bleach==6.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
bcrypt==4.1.2
email-validator
stripe
Werkzeug==3.0.1
python-dateutil==2.9.0
geoip2
user-agents
dnspython>=2.4.0
requests>=2.31.0
PyJWT>=2.8.0
